services restructured service in feature is found in dir feature/service service in shared is found in dir shared/services dir services and file data.service.ts deleted
