+++ src/app/login/login.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {FormGroup, FormBuilder, Validators} from '@angular/forms';
 import {User} from '../user/model/user';
-import { AuthService } from '../shared/auth/auth.service';
+import { AuthService } from '../shared/service/auth/auth.service';
 import {Router} from '@angular/router';
 
 @Component({
